add storage handler wrapper to accept different implementations of delete key method
